uses function to panic
